django/requirements.txt
Django==2.2
channels
docutils
requests
djangorestframework
coverage
